Ensure logging does not report spurious failures (#12265)
* Ensure logging does not report spurious failures

... and a bunch of other small improvements.

Previously, when shutting down a project in a normal way, user would be
spammed with
```
java.io.EOFException
	at java.base/java.io.ObjectInputStream$BlockDataInputStream.peekByte(ObjectInputStream.java:3232)
	at java.base/java.io.ObjectInputStream.readObject0(ObjectInputStream.java:1713)
	at java.base/java.io.ObjectInputStream.readObject(ObjectInputStream.java:540)
	at java.base/java.io.ObjectInputStream.readObject(ObjectInputStream.java:498)
	at org.enso.logging.service.logback.SocketLoggingNode.run(SocketLoggingNode.java:63)
	at java.base/java.lang.Thread.run(Thread.java:1583)
[DEBUG] [...] [org.enso.logging.service.logback.SocketLoggingNode] Caught java.io.EOFException closing connection.
```
That's because LoggingServer was unaware of LanguageServer being
stopped, or what client/project were actually connected with a socket.
This change ensures that no shutdown errors are being reported when in
shutdown mode.

In order to associate individual logging socket with a project, we make
use of Mapped Diagnostic Context (MDC) properties. Except that the info
was being lost during serialization. Had to provide a custom
`SocketAppender` that preserves MDC when event is being reported.

This has an additional advantage that MDC will be possible to be used
when logging events remotely and provide sufficient context for
indexing/search.

Added retries to `DummyRepository` as download tests were frequently
reporting false positives.

Added a single test to `NativeLibraryFinderTest` to investigate MacOS
failures.

* Reduce the amount of info logs + random failure

```
  pw:browser [pid=55591][err] Exception in thread "Logback SocketNode (client: /127.0.0.1:49521)" com.oracle.svm.core.jdk.UnsupportedFeatureError: SerializationConstructorAccessor class not found for declaringClass: java.util.Collections$UnmodifiableMap (targetConstructorClass: java.lang.Object). Usually adding java.util.Collections$UnmodifiableMap to serialization-config.json fixes the problem. +26s
  pw:browser [pid=55591][err] 	at org.graalvm.nativeimage.builder/com.oracle.svm.core.util.VMError.unsupportedFeature(VMError.java:121) +1ms
  pw:browser [pid=55591][err] 	at org.graalvm.nativeimage.builder/com.oracle.svm.core.reflect.serialize.SerializationSupport.getSerializationConstructorAccessor(SerializationSupport.java:163) +0ms
  pw:browser [pid=55591][err] 	at java.base@21.0.2/jdk.internal.reflect.MethodAccessorGenerator.generateSerializationConstructor(MethodAccessorGenerator.java:66) +0ms
  pw:browser [pid=55591][err] 	at java.base@21.0.2/jdk.internal.reflect.MethodAccessorGenerator.generateSerializationConstructor(MethodAccessorGenerator.java:54) +0ms
  pw:browser [pid=55591][err] 	at java.base@21.0.2/jdk.internal.reflect.ReflectionFactory.generateConstructor(ReflectionFactory.java:420) +0ms
  pw:browser [pid=55591][err] 	at java.base@21.0.2/jdk.internal.reflect.ReflectionFactory.newConstructorForSerialization(ReflectionFactory.java:412) +0ms
```
as reprted in https://github.com/enso-org/enso/actions/runs/13290451840/job/37117157855?pr=12265#step:13:77

* Potentially fix race condition between sbt tasks

By using `dependsOn` on SBT's tasks we define the correct order of
execution between tasks. Previously it wasn't guaranteed.

Closes #12273.

The interaction between native lib extraction and incremental
compilation wasn't playing well - StdBits automatic files update/removal
and directory cleaning was unpredictable for some of the stdlib
projects. I was able to reproduce CI issues locally.

Now, the behaviour is completely deterministic and predictable.