Ability to generate cloud credentials in the CLI build tool (#12438)
I needed to generate credentials for the staging environment.
The standard workflow to do this was to rebuild the IDE with a different `.env` file. However this took so long, as it was finishing (and the build failed for me), I was already finishing this script.

This should be useful for testing cloud integration from the CLI during development. It relies on the same logic as is used for generating credentials in our CI tests.

# Important Notes
Usage:
- You need to set up the same set of environment variables as for the CI workflow:
- `ENSO_CLOUD_COGNITO_REGION`, `ENSO_CLOUD_COGNITO_USER_POOL_ID`, `ENSO_CLOUD_COGNITO_USER_POOL_WEB_CLIENT_ID` - which can be taken from the `.env` file corresponding to a given environment, although there they have slightly different prefixes (as the config is for IDE not CI).
- `ENSO_CLOUD_TEST_ACCOUNT_USERNAME` and `ENSO_CLOUD_TEST_ACCOUNT_PASSWORD`
- Then running `./run backend generate-cloud-credentials` should create a `enso.credentials` file in your CWD. To make our libs use it, you can set `ENSO_CLOUD_CREDENTIALS_FILE` env var to point at it.