Emit a warning on non-AOT ready libraries (#12468)
The idea of this PR is to add a `Package.isAotReady()` flag. `EnsoLibraryFeature` is modified to _turn the flag on_ for every library package it processes - e.g. for all `Standard.*` currently. The flag is stored in a `PackageUtils.AOT_READY` static field of a class that is _initialized in build time_. As such all the flags `EnsoLibraryFeature` sets are visible during the execution in _native_ `enso` launcher.

# Important Notes
When @GregoryTravis separates the _generic JDBC driver_ into its own `Standard.Jdbc` library package, and such a package isn't going to be processed by `EnsoLibraryFeature`, then any project that `import Standard.Jdbc` will emit this warning.