Workaround ANTLR warnings in Native Image (#12486)
Truffle lists ANTLR as one of its dependencies and it appears that it takes a precedence over our version of ANTLR when building Native Image. Downgrading to the corresponding version to avoid annoying warnings.

Closes #12485.

Look Ma, there is even [ticket](https://github.com/oracle/graal/issues/5157) for it! But that workaround is not really applicable.

# Important Notes
This version will need to be updated every time GraalVM is updated.