Ensure correct BouncyCastle initialization for NI (#12822)
BouncyCastle hardly works out-of-the-box in NI. Instead one has to make sure that it is initialized at build time, with an exception of some classes that need to be done during runtime.
This change fixes key generation issues in `Auth_Spec`.

Cleaned up warnings related to wrong entries in nested enums:
```
...
[info] Warning: Could not register complete reflection metadata for org.openxmlformats.schemas.spreadsheetml.x2006.main.STSortMethod$Enum. Reason(s): java.lang.NoClassDefFoundError: org/openxmlformats/schemas/spreadsheetml/x2006/main/STSortMethod.
[info] Warning: Could not register complete reflection metadata for org.openxmlformats.schemas.spreadsheetml.x2006.main.STTableType$Enum. Reason(s): java.lang.NoClassDefFoundError: org/openxmlformats/schemas/spreadsheetml/x2006/main/STTableType.
...
```

Follow up  to #12804.  Closes #12813.

# Important Notes
Snowflake's and other tests appear to be fixed.