VSCode verifies `rootFolder` exists before using it (#12881)
Recently a change in `sbt` project structure (done as part of #12618) renamed a project and left a debris behind. One can remove it as
```
enso$ rm -r lib/java/desktop-environment/
```
and that fixes the problem as well. However this PR makes the Enso VSCode extension code more robust to "survive" such debris and not break other projects in the Git repository. Also it fixes the scan for _"all"_ projects.