Improve error for missing table ordering (#12917)
`add_row_number` for database backends requires an explicit or implicit table ordering, and if that is missing, the error states that `add_row_number` requires it. But other operations use `add_row_number` behind the scenes, such as `take`, and this error message is confusing because it mentions `add_row_number` which the user is probably not aware of using. This change adds an explicit check for `take` and `drop` and ensures that the error is attributed clearly to `take` or `drop`.

Closes #12785.