Instrument function parameters (#12936)
Attaching a visualization to a function's parameter was possible but yielded no value. This was because

a) the node was not instrumentable
b) had no location associated with it

both of which are fixed in this PR.
Additionally, if the parameter had a default value, the parser would assign an off-by-one location to it, preventing instrumentation and visualizations to work correctly.

Also fixed a warning caused by a recent GraalVM upgrade
```
WARNING: The 'graal.' property prefix for the Graal option PrintGraph
WARNING: (and all other Graal options) is deprecated and will be ignored
WARNING: in a future release. Please use 'jdk.graal.PrintGraph' instead.
```

Closes #11952.

# Important Notes
![Screenshot from 2025-04-23 13-56-51](https://github.com/user-attachments/assets/161bf057-280c-4dce-8d59-9894c28819ef)