Fix SQL `Vector_Builder` to work with deeper structures (#13030)
This fixes a Stack Overflow error occurring when running Snowflake tests.

The `Vector_Builder` that is used under the hood in the SQL builder was relying on non-tail recursion, meaning that if the structure to be built was nested too deeply, it could overflow the stack.

This PR addresses that by rewriting the `build` method to keep a stack on the heap, thus making the main loop tail-recursive. It adds a tests of `Vector_Builder` that verifies it can build very large structures of various shapes. The test case that builds 4 vectors that total 24000 elements finishes in ~0.5s on my machine.