Move Backend and Text providers to Vue (#13041)
**When reviewing, omit "d857c64beb178bd3736ffde29e3819987f9a73fd" (Use new useTextInReact) commit** - it's an automated work of replacing previous useText with useTextInReact.

When implementing #12977, I discovered that BackendProviderForVue does not work as expected - for some reason, its children were not rendered in same tree, so they didn't have access to provided contexts.

Therefore, I relied on an already proven method of sharing vue context in react instead, and moved the backend initialization to "vue" part. Because remote backend depended on `getText`, I moved this as well.