File browser extensions (#13048)
Closes #12004

Note that the extension lists are mocked in the videos below, so you will not see the same result in the app until the libs configuration is updated.

In cloud browser:

https://github.com/user-attachments/assets/be921752-3f75-4ff9-be33-1879f8fef708

In native browser:

https://github.com/user-attachments/assets/066c1e88-9bce-4dc5-8e1c-3973fbbf5948

Latest version (includes changes from the latest develop):

https://github.com/user-attachments/assets/d3c7a6d3-e8ad-4edb-967a-81e34f359f38

In read-mode file browser:

https://github.com/user-attachments/assets/de4cec3b-1f90-4014-aef6-d0931f643de7

# Important Notes
Interaction handler now supports nested interactions. I’m not 100% happy with the API, but I suggest gather more usage patterns before improving it if necessary.

Also, deep refactoring of the SelectionSubmenu widget, making it more universal.