Don't use `--enable-preview` option needlessly (#13204)
The usage of `--enable-preview` is causing problems when switching between frgaal and javac. JDK's compiler only allows `--enable-preview` with the latest `--target` e.g. `24`. This affects also other JDK utilities like `javadoc`. As a result I was unable to use `persistance/doc` to generate Javadoc when working on #13201.

The only reason why we are using`--enable-preview` is **usage of `_` in `switch` statements**. It is not really convincing reason for requiring preview mode. Hence this PR removes `--enable-preview` setting from `build.sbt` and changes the affected code to give the variable another name than `_`.

# Important Notes
The best solution would be for frgaal to `--enable-preview` by default (at the end the language is controlled by version of the frgaal compiler, no need for another switch) or offer some other (JDK javac unknown) option `-Denable.preview=true` to enable preview or offer an option to _disable preview_. That way it would be easier to switch between frgaal and latest JDK's `javadoc` without changing the arguments for the compilation. That's however a discussion to be held elsewhere.