Make socket appender tolerant to internal errors (#13605)
* Make socket appender tolerant to internal errors

When `SocketAppender` encounters errors during serialization of logging
events, it shutsdown the dispatcher. As a result, events are still being
logged (for some time) but no logs appear.
Under special circumstances thread starvation was also experienced.

As `AbstractSocketAppender` does not allow overriding dispatcher thread,
`DeferredProcessingSocketAppender` essentially copies its implementation
with few modifications. On `non-IOException` it wraps the error as a log
and reports.
As a result instead of silent logs we get a rather useful message with a
full stacktrace, e.g.,
```
[org.enso.interpreter.runtime.callable.UnresolvedSymbol] Internal error during serialization: Cannot invoke "org.enso.interpreter.runtime.EnsoContext.currentState()" because the return value of "org.enso.interpreter.runtime.EnsoContext.get(com.oracle.truffle.api.nodes.Node)" is null
```

* docs

* Update lib/scala/logging-service-logback/src/main/java/org/enso/logging/service/logback/DeferredProcessingSocketAppender.java

Co-authored-by: Jaroslav Tulach <jaroslav.tulach@enso.org>

---------

Co-authored-by: Jaroslav Tulach <jaroslav.tulach@enso.org>