Remove cache invalidation for visualizations (#13608)
A visualization or execute expression request always triggered cache invalidation for target expression. This was meant to allow re-calculation for subexpressios that are part of bigger (cached) expressions.
It also can cause massive delays when expressions are computationally-intensive.

However cache-invalidation is currently unnecessary, as idmap is set before first execution and we currently don't attach visualizations to self arguments (that may change in the future).
The better way to handle this is to make cache reactive, which is the role of #10525. That would make cache invalidation logic obsolete.

This change improves the overall startup time as it reduces constant re-compilation of expensive nodes.

Improves #13580. The attempt was previously attempted in https://github.com/enso-org/enso/pull/11949 but reverted due to widgets regression.