cleanup svg icon usage, types and href generation (#13746)
Fixes #13712

Removed remains of svg embedding, as it turned out to not work reliably when combined with shadow-dom (i.e. in visualizations) and the exact behavior is browser dependent. I ended up just unifying all places that reference icon's `href` into a single function, and added a "missing icon" visual. Unfortunately that means the playwright test snapshots still attempt a potentially cross-origin fetch request to resolve the icons, which is cancelled before hitting playwright's service-worker fetch handler. Allowing cross-origin requests in svg is often possible, but for `<use>` tags is not yet implemented anywhere. The only workaround that somewhat works is to run the test reporter on the same port as the app under test used to have. We unfortunately use two, so the icons will either show up in dashboard or in project-view tests.

Also deduplicated login logic in playwright tests and added extra agreement modal detection, since that modal often used to be a problem for me when running test suite locally.