Forbid function expressions in statement context: Implementation (#13920)
Implement a syntax error when a function-creating expression (e.g. an operator section, template function, or lambda) is used as a statement--doing so is never useful, and likely to be a mistake.

This PR adds support for detecting the case to the parser; the error is not enabled yet, as #13955 must be resolved first.

# Important Notes
- This is a breaking language change: Certain types of dead code, that were previously ignored, are no longer allowed.
- Note that this includes `//`-style "comments", which are no longer allowed in Enso code. They used to sometimes "work", but now will always trigger the suspicious-dead-code error.
- Also: Ported syntax error tests to `insta`, and extended them to cover error messages and locations.
- All syntax errors now are now defined as `SyntaxError` values. This improves maintainability, and enables more efficient parser FFI.
- More consistent syntax error punctuation.