Remove content reflow when resizing right panel (#13976)
Closes #13711

Also fixed an issue when the right panel width was not contained within the screen size.
I also changed the colors of the right panel to pure white — it matches the dashboard now and has no weird ”step“ between the inner content background and the panel background.

Develop, 10x speed of animation:

https://github.com/user-attachments/assets/dae12fe6-5dae-4b2c-b2e3-d4dce1385c0d

Now, 10x speed of animation:

https://github.com/user-attachments/assets/f96788ce-1a24-4cbd-8920-d074545828be

Realtime:

https://github.com/user-attachments/assets/cc84f462-e539-4c93-9f90-65b84790d2f0