Fix blank arguments in signature (#14078)
Signature generator prints blank (underscore) arguments, instead of `<internal-*>`.

On develop, signature for
```
func _ =
42
```
is `func <internal-0> -> Standard.Base.Any.Any`, on this PR, the signature is `func _ -> Standard.Base.Any.Any`

# Important Notes
The compiler transform all blank (underscore) arguments into arguments with `<internal-\d+>` literal names. The information that the argument was created from an underscore is lost in the compilation pipeline. This PR adds a workaround to signature generator: if such argument is encountered, simple `_` is printed instead of its internal name.