Move Upload File to Cloud logic to Vue (#14107)
A prelude to #13491

Move upload file logic to Vue. It will be needed when handling hybrid projects.

# Important Notes
Because of various problems with `tanstack` query we faced lately, which were hard to debug, I decided to _not_ use it in the new implementation. I moved the "pooling" mechanism to the new store.