Fix dispatching events unique to pre event registry (#14186)
Fixed a bug with visualization preview being "sticky" - releasing ctrl not causing it to immediately close, but required another keypress.

The global event registry had a bug that caused events registered only in `pre` handlers to not fire (e.g. `keyup` event in global keyboard state). This was caused by the events always having to originate from non-pre registry, but it not having any handlers actually registered. Now the `pre` and `post` handlers are treated as two dispatch phases in the same registry.