Replace synchronization blocks in polyglot Java initialization (#13907)
* Replace synchronization blocks in polyglot Java initialization

If 2 threads are trying to initialize polyglot Java
- Thread 1 acquires the lock on the object and proceeds with
 initialization
- Thread 2 waits on the lock on the object

Sadly, Thread 1, while keeping the lock hits the TruffleSafepoint
barrier, waiting for all other threads to poll at least once.
Thread 2 will never call TruffleSafepoint.poll() as it is waiting.

This change replaces synchronization block with a lightweight access to
a sempahore. Single cell semaphore lock is attempted with a `tryAcquire`
call until successful. `acquire` must not be used as it would lead to
the problem described above again.
Busy polling for a sempahore is being used. It seems acceptable as
creating a language is not a resource-intensive process and based on
local testing proceeds quickly with little wasted resource.

* Allow core threads to timeout in guest executor

Core threads waiting on new work interere with Truffle's safepoint
polling (they never call it). This change allows for core threads to
timeout which eventually allow the truffle barrier to move on, if
encountered.

* Missed one synchronization during rebase

* Fix for a problem introduced during rebase

* Test demonstrating multithreaded lang initialization

Test would previously get stuck or rather timeout, due to lack of
progress.

Also discovered and fixed that initialization of Extra values was not thread-safe
- was getting random assertion failures in tests.

* fmt

* Replace busy waiting with TruffleSafepoint construct

Wrapping `Semaphore.acquire` in
`TruffleSafepoint.setBlockedThreadInterruptible` apparently makes it
Truffle friendly and there is no need for busy waiting. Profit.

* Semaphore is interruptable only when context is entered

Calling `TruffleSafepoint.setBlockedThreadInterruptible` on an object
prior to entering a polyglot context would result in cryptic errors.
Sadly, we do actually add files to classpath during compilation. By
falling back to plain `Semaphore::acquire` call, Truffle's asserts are
not triggered.

* print all tests to debug CI issues

* bump max IR size

* (debugging) Check vm.max_map_count

* Testing bumped vm.max_map_count value

* Revert max_map_count changes

We are running std lib native tests on self-hosted runners. Changes must
be made directly on the machines.
Was able to reproduce the issue locally with a very small value.

---------

Co-authored-by: Jaroslav Tulach <jaroslav.tulach@enso.org>