Distinguish property access in AST (#14226)
Add a `PropertyAccess` node to the Rust AST, representing use of the dot `.` operator. This operator was previously treated as a special case both in the GUI and in the backend.

The `PropertyAccess` node only represents syntactically-valid applications of the operator. Its RHS is non-optional, and is an identifier token, not an expression.

This eliminates the largest difference between AST and Tree, and so is a major step toward #10754.