Ignore MacOS' auxilary directories (#14369)
It appears that on some MacOS CI runners some indexing is turned on,
which means that OS creates automatically hidden directories.
This messes up with our expectations. This change adds an exception
for MacOS to simply ignore `._` prefixed directories during unpacking
in `Archive.scala`.

Simpler version of #14340 which tried to fix dependencies as well but
instead opened a can of worms.
Closes #14294.